+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2019-08-31
+
+### Added
+
+- Comments History
+- Mute and Unmute Comments
+
 ## [1.2.0] - 2019-08-19
 
 ### Added